requirements.txt
# Core runtime
pydantic==2.7.4
httpx==0.27.0
playwright==1.48.0
rich==13.9.2
ujson==5.10.0
pyyaml==6.0.2
pandas>=2.0
requests>=2.30

# Discovery & scraping helpers
duckduckgo-search>=6.1.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Enrichment providers (optional but enabled for you)
upgini>=1.1

# Fast JSON (optional; our logger uses it if present)
orjson>=3.10.0 ; platform_machine != "i686" and platform_python_implementation == "CPython"
